refactor♻️: Realocado arquivos de testes para uma pasta separada, ao padrao pythonico
